refactor content stucture
