+++ front_end/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
-import 'package:front_end/api/firebase_api.dart';
+import 'package:front_end/util/firebase_messaging.dart';
 import 'package:front_end/ui/root_page/root_page.dart';
 
 Future<void> main() async {
